+++ packages/cli/doc-plugin-api-docgen/src/docgen.ts
@@ -1,6 +1,6 @@
 import path from 'path';
-import { logger } from '@modern-js/utils';
-import type { ComponentDoc } from 'react-docgen-typescript';
+import type { ComponentDoc, PropItem } from 'react-docgen-typescript';
+import { logger, chokidar, fs } from '@modern-js/utils';
 import { parse } from 'react-docgen-typescript';
 import { apiDocMap } from './constants';
 import { locales } from './locales';
@@ -9,6 +9,7 @@ import type {
   Entries,
   ToolEntries,
   ApiParseTool,
+  WatchFileInfo,
 } from './types';
 
 const isToolEntries = (obj: Record<string, any>): obj is ToolEntries => {
@@ -20,7 +21,10 @@ export const docgen = async ({
   languages,
   apiParseTool,
   appDir,
+  parseToolOptions,
+  isProd,
 }: DocGenOptions) => {
+  const watchFileMap: Record<string, WatchFileInfo> = {};
   const genApiDoc = async (entry: Entries, tool: ApiParseTool) => {
     if (Object.keys(entry).length === 0) {
       return;
@@ -28,17 +32,46 @@ export const docgen = async ({
     await Promise.all(
       Object.entries(entry).map(async ([key, value]) => {
         const moduleSourceFilePath = path.resolve(appDir, value);
+        watchFileMap[moduleSourceFilePath] = {
+          apiParseTool,
+          moduleName: key,
+        };
         try {
           if (tool === 'documentation') {
             const documentation = await import('documentation');
 
-            const documentationRes = await documentation.build([
-              moduleSourceFilePath,
-            ]);
-            const apiDoc = await documentation.formats.md(documentationRes);
+            const documentationRes = await documentation.build(
+              [moduleSourceFilePath],
+              {
+                ...parseToolOptions.documentation,
+              },
+            );
+            const apiDoc = await documentation.formats.md(documentationRes, {
+              noReferenceLinks:
+                parseToolOptions.documentation?.noReferenceLinks ?? true,
+            });
             apiDocMap[key] = apiDoc;
           } else {
-            const componentDoc = parse(moduleSourceFilePath);
+            const componentDoc = parse(moduleSourceFilePath, {
+              // https://github.com/styleguidist/react-docgen-typescript/blob/master/README.md?plain=1#L111
+              propFilter: (prop: PropItem) => {
+                if (
+                  prop.declarations !== undefined &&
+                  prop.declarations.length > 0
+                ) {
+                  const hasPropAdditionalDescription = prop.declarations.find(
+                    declaration => {
+                      return !declaration.fileName.includes('node_modules');
+                    },
+                  );
+
+                  return Boolean(hasPropAdditionalDescription);
+                }
+
+                return true;
+              },
+              ...parseToolOptions['react-docgen-typescript'],
+            });
             if (componentDoc.length === 0) {
               logger.warn(
                 '[module-doc-plugin]',
@@ -74,6 +107,47 @@ export const docgen = async ({
     ]);
   } else {
     await genApiDoc(entries, apiParseTool);
+  }
+  if (!isProd) {
+    const watcher = chokidar.watch(Object.keys(watchFileMap), {
+      ignoreInitial: true,
+      ignorePermissionErrors: true,
+      ignored: [/node_modules/],
+    });
+    let isUpdate = false;
+    watcher.on('change', changed => {
+      if (isUpdate) {
+        return;
+      }
+      isUpdate = true;
+      logger.info('[module-doc-plugin]', 'updating API');
+      const watchFileInfo = watchFileMap[changed];
+      if (watchFileInfo) {
+        const { apiParseTool, moduleName } = watchFileInfo;
+
+        const updateSiteData = () => {
+          const siteDataPath = path.join(
+            process.cwd(),
+            'node_modules',
+            '.modern-doc',
+            'runtime',
+            'virtual-site-data.mjs',
+          );
+          import(siteDataPath).then(siteData => {
+            const data = { ...siteData.default };
+            data.pages.forEach((page: Record<string, unknown>) => {
+              page.apiDocMap = apiDocMap;
+            });
+            fs.writeFileSync(
+              siteDataPath,
+              `export default ${JSON.stringify(data)}`,
+            );
+            isUpdate = false;
+          });
+        };
+        genApiDoc({ [moduleName]: changed }, apiParseTool).then(updateSiteData);
+      }
+    });
   }
 
   logger.success('[module-doc-plugin]', `Generate API table successfully!`);