+++ packages/cli/doc-plugin-api-docgen/src/index.ts
@@ -17,10 +17,11 @@ export function pluginApiDocgen(options?: PluginOptions): DocPlugin {
     entries = {},
     apiParseTool = 'react-docgen-typescript',
     appDir = process.cwd(),
+    parseToolOptions = {},
   } = options || {};
   return {
     name: '@modern-js/doc-plugin-api-docgen',
-    async beforeBuild(config) {
+    async beforeBuild(config, isProd) {
       // only support zh and en
       const languages = (
         config.themeConfig?.locales?.map(locale => locale.lang) ||
@@ -31,6 +32,8 @@ export function pluginApiDocgen(options?: PluginOptions): DocPlugin {
         apiParseTool,
         languages,
         appDir,
+        parseToolOptions,
+        isProd,
       });
     },
     async modifySearchIndexData(pages) {