+++ packages/server/babel-plugin-module-resolver/.eslintrc.js
@@ -1,0 +1,18 @@
+module.exports = {
+  root: true,
+  extends: ['@modern-js'],
+  rules: {
+    'import/named': 'off',
+    'import/namespace': 'off',
+    'import/default': 'off',
+    'import/no-named-as-default-member': 'off',
+    'import/no-named-as-default': 'off',
+    'import/no-cycle': 'off',
+    'import/no-unused-modules': 'off',
+    'import/no-deprecated': 'off',
+  },
+  parserOptions: {
+    tsconfigRootDir: __dirname,
+    project: ['./tsconfig.json'],
+  },
+};