+++ packages/server/babel-plugin-module-resolver/src/log.js
@@ -1,0 +1,5 @@
+// This module exists only for abstracting logging away and making testing easier
+
+export function warn(...args) {
+  console.warn(...args);
+}