+++ tests/e2e/builder/cases/vue2/jsx-basic/src/index.js
@@ -1,0 +1,8 @@
+import Vue from 'vue';
+import A from './A';
+
+// eslint-disable-next-line no-new
+new Vue({
+  el: '#root',
+  render: h => h(A),
+});