+++ tests/e2e/builder/cases/vue2/sfc-style/src/index.js
@@ -1,0 +1,9 @@
+import Vue from 'vue';
+import App from './App.vue';
+import './a.less';
+
+// eslint-disable-next-line no-new
+new Vue({
+  el: '#root',
+  render: h => h(App),
+});