+++ packages/builder/builder-rspack-provider/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @modern-js/builder-rspack-provider
 
+## 2.27.0
+
+### Minor Changes
+
+- 3f79dd7: feat(builder): add `source.aliasStrategy` option
+
+  feat(builder): 新增 `source.aliasStrategy` 选项
+
+### Patch Changes
+
+- 645e111: fix(builder): CSS Modules name has a probability of conflict in prod
+
+  fix(builder): 修复 CSS Modules name 在生产环境有极小概率冲突的问题
+
+- 8322a51: chore: migrate packages from tsc to module-lib build
+
+  chore: 将使用 tsc 的包迁移到 module-lib 构建
+
+- 5376a22: chore(builder): update rspack to 0.2.8 and compatible [dev-client change](https://github.com/web-infra-dev/rspack/pull/3731/files)
+
+  chore(builder): 升级 rspack 到 0.2.8 并兼容 [dev-client 变更](https://github.com/web-infra-dev/rspack/pull/3731/files)
+
+- 67d0b0c: fix(builder): failed to disable html via htmlPlugin: false
+
+  fix(builder): 修复通过 htmlPlugin: false 无法禁用 html 的问题
+
+- Updated dependencies [645e111]
+- Updated dependencies [91d14b8]
+- Updated dependencies [8322a51]
+- Updated dependencies [d9080ed]
+- Updated dependencies [67d0b0c]
+- Updated dependencies [3f79dd7]
+- Updated dependencies [6d7104d]
+  - @modern-js/builder-shared@2.27.0
+  - @modern-js/utils@2.27.0
+  - @modern-js/server@2.27.0
+  - @modern-js/types@2.27.0
+
 ## 2.26.0
 
 ### Patch Changes