+++ packages/generator/generators/module-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @modern-js/module-generator
 
+## 3.1.29
+
 ## 3.1.28
 
 ## 3.1.27