+++ packages/server/server/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/server
 
+## 2.27.0
+
+### Patch Changes
+
+- Updated dependencies [91d14b8]
+- Updated dependencies [6d7104d]
+  - @modern-js/utils@2.27.0
+  - @modern-js/prod-server@2.27.0
+  - @modern-js/server-utils@2.27.0
+  - @modern-js/types@2.27.0
+
 ## 2.26.0
 
 ### Patch Changes