+++ packages/server/utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @modern-js/server-utils
 
+## 2.27.0
+
+### Patch Changes
+
+- Updated dependencies [91d14b8]
+- Updated dependencies [8322a51]
+- Updated dependencies [6d7104d]
+  - @modern-js/utils@2.27.0
+  - @modern-js/babel-compiler@2.27.0
+  - @modern-js/babel-preset-base@2.27.0
+  - @modern-js/babel-plugin-module-resolver@2.27.0
+
 ## 2.26.0
 
 ### Patch Changes